font: use custom font
Created a new font file with icons for nakamochi.

Renamed font file to use generic filename that only includes size
information so that the code does not have to be changed if the
font is replaced with another font style.

Note: this font is not yet production ready, need to create a
test view in ngui to render and compare all icons side by side.