ui: force alignment of modal callback function
fullscreen modal dialog was added in 4297c139. the callback function to
that modal *must* be aligned according to the target architecture.
failing to do so may result in "incorrect alignment" panic.

this somehow worked in 4297c139 but sometimes panic. i suspect this may
be due to PIE-enabled build, possibly in conjuction with some zig v0.10
compiler bugs.

ref https://git.qcode.ch/nakamochi/ndg/pulls/20
ref https://git.qcode.ch/nakamochi/ndg/issues/5