ui/driver: add Xlib support when dev-building on a linux host
something broken in recent SDL2 or Xorg releases, or a combination of
plus LVGL usage of their APIs. the result was that the main window
stopped drawing altogether. Downgrading SDL2 and other parts didn't seem
to help at all.

this commit adds support for SDL2-alternative when compiling and running
the GUI on a dev linux machine with Xorg.

there's a small patch to the original lv_drivers/x11/x11.c since clang
in C11 mode refuses to compile otherwise, giving a "initializer element
is not a compile-time constant" error.

release builds remain as is. this is only for development.